MapSchema.Examples.Person (map_schema v0.1.0)

Functions

get_age(arg)

Get age value of integer type.

get_contact_email(arg)

Get contact_email value of string type.

get_contact_others(arg)

Get contact_others value of any type.

get_contact_phone(arg)

Get contact_phone value of string type.

get_country(arg)

Get country value of string type.

get_name(arg)

Get name value of string type.

get_surname(arg)

Get surname value of string type.

is_valid?(arg)

It´s will return true, if the map follow the schema.

json_decode(arg)

Let decode json to Object. Checking every type following the schema.

Note: This method use Jason library.

json_decode(arg1, arg2)

Specs

json_decode(mapa :: MapSchema.Examples.Person, json :: BitString) :: Any

Let decode json and mut a existing object . Checking every type following the schema. Intenal it´s using the method put/2.

Note: This method use Jason library.

Parameters

mapa: Object

json: Json object

json_encode(arg)

Let encode object to Json.

Note: This method use Jason library.

mut_age(arg1, arg2)

Change the value of age field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that age is integer type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_age(fn(_old)-> 20 end)
  iex> Elixir.MapSchema.Examples.Person.get_age(obj)
  20

mut_contact_email(arg1, arg2)

Change the value of contact_email field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that contact_email is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_contact_email(fn(_old)-> "Barcelona" end)
  iex> Elixir.MapSchema.Examples.Person.get_contact_email(obj)
  "Barcelona"

mut_contact_others(arg1, arg2)

Change the value of contact_others field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that contact_others is any type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_contact_others(fn(_old)-> "its_anything" end)
  iex> Elixir.MapSchema.Examples.Person.get_contact_others(obj)
  "its_anything"

mut_contact_phone(arg1, arg2)

Change the value of contact_phone field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that contact_phone is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_contact_phone(fn(_old)-> "Galicia" end)
  iex> Elixir.MapSchema.Examples.Person.get_contact_phone(obj)
  "Galicia"

mut_country(arg1, arg2)

Change the value of country field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that country is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_country(fn(_old)-> "Galicia" end)
  iex> Elixir.MapSchema.Examples.Person.get_country(obj)
  "Galicia"

mut_name(arg1, arg2)

Change the value of name field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that name is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_name(fn(_old)-> "Madrid" end)
  iex> Elixir.MapSchema.Examples.Person.get_name(obj)
  "Madrid"

mut_surname(arg1, arg2)

Change the value of surname field, using a mut_fn.(actualValue) that will recive the actual value. Before of update the valor will be check it.

Remember that surname is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.mut_surname(fn(_old)-> "Barcelona" end)
  iex> Elixir.MapSchema.Examples.Person.get_surname(obj)
  "Barcelona"

new()

This method only create a new simple map %{}. Yes.

Because the idea it´s the data structure will be independent of MapSchema but the module will have the schema that the maps should be follow.

Example:

  iex>Elixir.MapSchema.Examples.Person.new()
  %{}

put(arg)

Put a new value in each field of the update map, in the field of the object.

But before of update the values always will be check the type.

put(arg1, arg2)

Put a new value in each field of the update map, in the field of the object.

But before of update the values always will be check the type.

put_age(arg1, arg2)

Put a new value in age field. Before of update the valor will be check it.

Remember that age is integer type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_age(50)
  iex> Elixir.MapSchema.Examples.Person.get_age(obj)
  50

put_contact_email(arg1, arg2)

Put a new value in contact_email field. Before of update the valor will be check it.

Remember that contact_email is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_contact_email("Barcelona")
  iex> Elixir.MapSchema.Examples.Person.get_contact_email(obj)
  "Barcelona"

put_contact_others(arg1, arg2)

Put a new value in contact_others field.

put_contact_phone(arg1, arg2)

Put a new value in contact_phone field. Before of update the valor will be check it.

Remember that contact_phone is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_contact_phone("Sevilla")
  iex> Elixir.MapSchema.Examples.Person.get_contact_phone(obj)
  "Sevilla"

put_country(arg1, arg2)

Put a new value in country field. Before of update the valor will be check it.

Remember that country is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_country("Sevilla")
  iex> Elixir.MapSchema.Examples.Person.get_country(obj)
  "Sevilla"

put_name(arg1, arg2)

Put a new value in name field. Before of update the valor will be check it.

Remember that name is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_name("Santiago")
  iex> Elixir.MapSchema.Examples.Person.get_name(obj)
  "Santiago"

put_surname(arg1, arg2)

Put a new value in surname field. Before of update the valor will be check it.

Remember that surname is string type.

Example:

  iex>obj = Elixir.MapSchema.Examples.Person.new()
  ...> |> Elixir.MapSchema.Examples.Person.put_surname("Madrid")
  iex> Elixir.MapSchema.Examples.Person.get_surname(obj)
  "Madrid"

schema()

The Schema is: %{"age" => :integer, "contact" => %{"email" => :string, "others" => :any, "phone" => :string}, "country" => :string, "name" => :string, "surname" => :string}